only delete Symbology if present

+++ R/get_nlum.R
@@ -130,16 +130,17 @@
           fs::path(download_dir, "NLUM_v7_DescriptiveMetadata_20241128_0.pdf")
         )
       }
-
-      fs::dir_delete(
-        c(fs::path(nlum_dir, "Maps"), fs::path(nlum_dir, "Symbology"))
-      )
-      fs::file_delete(
-        setdiff(
-          fs::dir_ls(nlum_dir),
-          fs::dir_ls(nlum_dir, regexp = "[.]tif$|[.]tif[.]aux[.]xml$")
+      if (startswith(.data_set, "Y") || startswith(.data_set, "C")) {
+        fs::dir_delete(
+          c(fs::path(nlum_dir, "Maps"), fs::path(nlum_dir, "Symbology"))
         )
-      )
+        fs::file_delete(
+          setdiff(
+            fs::dir_ls(nlum_dir),
+            fs::dir_ls(nlum_dir, regexp = "[.]tif$|[.]tif[.]aux[.]xml$")
+          )
+        )
+      }
     },
     error = function(e) {
       cli::cli_abort(
